Update R-CMD-check to be based off r-lib actions exampls
The main change here is to expand the number of R versions tested. It will now test more versions but there is an argument 'must_pass' indicating whether the overall workflow will pass or fail if that job fails.

The other changes just move to using the r-lib pre-built actions which look to do pretty much the same thing as the exisiting custom ones but should be easier to maintain.