Expanding the CI workflows (#48)
* Update R-CMD-check to be based off r-lib actions exampls

The main change here is to expand the number of R versions tested. It will now test more versions but there is an argument 'must_pass' indicating whether the overall workflow will pass or fail if that job fails.

The other changes just move to using the r-lib pre-built actions which look to do pretty much the same thing as the exisiting custom ones but should be easier to maintain.

* Add workflow to document the package

This will run whenever a file in the `R` directory is modified.

* Add workflow to style code with `styler`

This uses the tidyverse style as default and will automatically make cosmetic changes to the code whenever an appropriate file type is modified.

* Add a workflow to lint the packge with `lintr`

This will run on any change to main/master and will add notes to any PR / commit with lint suggestions.